Rename plugin API package
